[feat] 06/28 프로필 등록 단계에서도 프로필 유무 확인 후 있을 경우 메인 페이지로 리다이렉트 #28
